+++ src/Frontend/Core/Engine/TemplateModifiers.php
@@ -9,6 +9,7 @@ namespace Frontend\Core\Engine;
  * file that was distributed with this source code.
  */
 
+use Frontend\Core\Engine\Model as FrontendModel;
 use Frontend\Core\Engine\Block\Widget as FrontendBlockWidget;
 use Frontend\Modules\Profiles\Engine\Model as FrontendProfilesModel;
 
@@ -375,7 +376,17 @@ class TemplateModifiers
      */
     public static function parseWidget($var, $module, $action, $id = null)
     {
-        $data = $id !== null ? serialize(array('id' => $id)) : null;
+        // we have an id
+        if($id !== null) {
+            // we should get the serialized data we don't have yet
+            $data = (string) FrontendModel::getContainer()->get('database')->getVar(
+                'SELECT i.data
+                 FROM modules_extras AS i
+                 WHERE i.module = ? AND i.action = ? AND i.id = ?',
+                array($module, $action, $id)
+            );
+        // we don't have an id
+        } else $data = null;
 
         // create new widget instance and return parsed content
         $extra = new FrontendBlockWidget(Model::getKernel(), $module, $action, $data);